Support for collecting images, flavors, and networks with OpenStack infra provider
Refactored common code out of OpenStack cloud parser into a mixin
so that it can be shared with the infra provider.

Flavors and networks collection does not work at the moment.
They depend on modeling changes to EmsOpenstackInfra which will
be done in other patches.
