+++ vmdb/app/models/ems_refresh/parsers/openstack.rb
@@ -2,6 +2,7 @@
 
 module EmsRefresh::Parsers
   class Openstack < Cloud
+    include OpenstackMixin
     # Openstack uses numbers to represent different power states. Each openstack
     # power state value corresponds to an array index for the human readable
     # power state.
@@ -70,22 +71,8 @@ module EmsRefresh::Parsers
       @security_groups ||= @network_service.security_groups_for_accessible_tenants
     end
 
-    def networks
-      @networks ||= @network_service.networks
-    end
-
     def volumes
       @volumes ||= @volume_service.volumes_for_accessible_tenants
-    end
-
-    def get_flavors
-      flavors = @connection.flavors
-      process_collection(flavors, :flavors) { |flavor| parse_flavor(flavor) }
-    end
-
-    def get_private_flavor(id)
-      private_flavor = @connection.flavors.get(id)
-      process_collection([private_flavor], :flavors) { |flavor| parse_flavor(flavor) }
     end
 
     def get_availability_zones
@@ -126,22 +113,6 @@ module EmsRefresh::Parsers
       end
     end
 
-    def get_networks
-      return unless @network_service_name == :neutron
-
-      process_collection(networks, :cloud_networks) { |n| parse_network(n) }
-      get_subnets
-    end
-
-    def get_subnets
-      return unless @network_service_name == :neutron
-
-      networks.each do |n|
-        new_net = @data_index.fetch_path(:cloud_networks, n.id)
-        new_net[:cloud_subnets] = n.subnets.collect { |s| parse_subnet(s) }
-      end
-    end
-
     # def get_hosts
     #   hosts = @connection.hosts.select { |h| h.service == "compute" }
     #   process_collection(hosts, :hosts) { |host| parse_host(host) }
@@ -168,11 +139,6 @@ module EmsRefresh::Parsers
           process_collection(fd.files, :cloud_object_store_objects) { |o| parse_object(o, result, t) }
         end
       end
-    end
-
-    def get_images
-      images = @image_service.images_for_accessible_tenants
-      process_collection(images, :vms) { |image| parse_image(image) }
     end
 
     def get_servers
@@ -226,26 +192,6 @@ module EmsRefresh::Parsers
         base_snapshot = @data_index.fetch_path(:cloud_volume_snapshots, base_snapshot_uid)
         cv[:base_snapshot] = base_snapshot unless base_snapshot.nil?
       end if @data[:cloud_volumes]
-    end
-
-    def parse_flavor(flavor)
-      uid = flavor.id
-
-      new_result = {
-        :type    => "FlavorOpenstack",
-        :ems_ref => uid,
-        :name    => flavor.name,
-        :enabled => !flavor.disabled,
-        :cpus    => flavor.vcpus,
-        :memory  => flavor.ram.megabytes,
-
-        # Extra keys
-        :root_disk      => flavor.disk.to_i.gigabytes,
-        :ephemeral_disk => flavor.ephemeral.to_i.gigabytes,
-        :swap_disk      => flavor.swap.to_i.megabytes
-      }
-
-      return uid, new_result
     end
 
     def parse_availability_zone(az)
@@ -365,31 +311,6 @@ module EmsRefresh::Parsers
       }
     end
 
-    def parse_network(network)
-      uid     = network.id
-      status  = (network.status.to_s.downcase == "active") ? "active" : "inactive"
-
-      new_result = {
-        :name            => network.name,
-        :ems_ref         => uid,
-        :status          => status,
-        :enabled         => network.admin_state_up,
-        :external_facing => network.router_external,
-        :cloud_tenant    => @data_index.fetch_path(:cloud_tenants, network.tenant_id)
-      }
-      return uid, new_result
-    end
-
-    def parse_subnet(subnet)
-      {
-        :name             => subnet.name,
-        :ems_ref          => subnet.id,
-        :cidr             => subnet.cidr,
-        :network_protocol => "ipv#{subnet.ip_version}",
-        :gateway          => subnet.gateway_ip,
-        :dhcp_enabled     => subnet.enable_dhcp,
-      }
-    end
 
     def parse_volume(volume)
       log_header = "MIQ(#{self.class.name}.#{__method__})"
@@ -474,32 +395,6 @@ module EmsRefresh::Parsers
         :tenant         => @data_index.fetch_path(:cloud_tenants, tenant.id)
       }
       return uid, new_result
-    end
-
-    def parse_image(image)
-      uid = image.id
-
-      parent_server_uid = parse_image_parent_id(image)
-
-      new_result = {
-        :type            => "TemplateOpenstack",
-        :uid_ems         => uid,
-        :ems_ref         => uid,
-        :name            => image.name,
-        :vendor          => "openstack",
-        :raw_power_state => "never",
-        :template        => true,
-        :publicly_available => image.is_public,
-      }
-      new_result[:parent_vm_uid] = parent_server_uid unless parent_server_uid.nil?
-      new_result[:cloud_tenant]  = @data_index.fetch_path(:cloud_tenants, image.owner) if image.owner
-
-      return uid, new_result
-    end
-
-    def parse_image_parent_id(image)
-      image_parent = @image_service_name == :glance ? image.copy_from : image.server
-      image_parent["id"] if image_parent
     end
 
     def parse_server(server)
